test meeting authorizations: extract to own module
